backend/requirements.txt
django~=5.2
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-decouple>=3.8
fastapi>=0.109
uvicorn[standard]>=0.27
celery>=5.3
redis>=5.0
thefuzz>=0.22
python-Levenshtein>=0.25
crewai>=0.28
crewai-tools>=0.4
openai>=1.12
openpyxl>=3.1
pandas>=2.00
python-dotenv>=1.0
pydantic>=2.5
Pillow>=10.2
djangorestframework-simplejwt>=5.3
